Reactive Item API.
Fixed the failure of tests during build. ConnectionFactoryInitializer does seem to be called for every test only for complete build.
Hence had to deal with dropping table if already exists. Going forward, we may have to move to Liquibase to make this table creation easy.
Auto table creation doe not seem to happen for R2DBC.